+++ src/coding/mpeg_decoder.h
@@ -7,27 +7,26 @@
 
 /* used by mpeg_decoder.c, but scattered in other .c files */
 #ifdef VGM_USE_MPEG
-#include <mpg123.h>
 
 /* represents a single MPEG stream */
 typedef struct {
-    /* per stream as sometimes mpg123 must be fed in passes if data is big enough (ex. EALayer3 multichannel) */
-    uint8_t* buffer; /* raw data buffer */
+    /* buf per stream as sometimes mpg123 must be fed in passes if data is big enough (ex. EALayer3 multichannel) */
+    uint8_t* buffer;
     size_t buffer_size;
     size_t bytes_in_buffer;
-    bool buffer_full; /* raw buffer has been filled */
-    bool buffer_used; /* raw buffer has been fed to the decoder */
+    bool buffer_full;
+    bool buffer_used;
 
-    mpg123_handle* handle; /* MPEG decoder */
+    void* handle;
 
-    void* sbuf; /* decoded samples from this stream */
-    size_t sbuf_size; /* in bytes for mpg123 */
-    size_t samples_filled; /* data in the buffer (in samples) */
-    size_t samples_used; /* data extracted from the buffer */
+    int16_t* sbuf;
+    int sbuf_size;      // in bytes for mpg123
+    int samples_filled;
+    int samples_used;
 
-    size_t current_size_count; /* data read (if the parser needs to know) */
-    size_t current_size_target; /* max data, until something happens */
-    size_t decode_to_discard;  /* discard from this stream only (for EALayer3 or AWC) */
+    int current_size_count; /* data read (if the parser needs to know) */
+    int current_size_target; /* max data, until something happens */
+    int decode_to_discard;  /* discard from this stream only (for EALayer3 or AWC) */
 
     int channels_per_frame; /* for rare cases that streams don't share this */
 } mpeg_custom_stream;
@@ -35,20 +34,20 @@ typedef struct {
 struct mpeg_codec_data {
     /* regular/single MPEG internals */
     uint8_t* buffer; /* raw data buffer */
-    size_t buffer_size;
-    size_t bytes_in_buffer;
+    int buffer_size;
+    int bytes_in_buffer;
     bool buffer_full; /* raw buffer has been filled */
     bool buffer_used; /* raw buffer has been fed to the decoder */
 
-    mpg123_handle* m; /* MPEG decoder */
-    struct mpg123_frameinfo mi; /* start info, so it's available even when resetting */
+    void* handle; // MPEG decoder
 
     /* for internal use */
     int channels_per_frame;
     int samples_per_frame;
     /* for some calcs */
-    int bitrate_per_frame;
-    int sample_rate_per_frame;
+    int bitrate;
+    int sample_rate;
+    bool is_vbr;
 
     /* custom MPEG internals */
     bool custom; /* flag */
@@ -57,11 +56,13 @@ struct mpeg_codec_data {
 
     size_t default_buffer_size;
     mpeg_custom_stream* streams; /* array of MPEG streams (ex. 2ch+2ch) */
-    size_t streams_size;
+    int streams_count;
 
-    size_t skip_samples; /* base encoder delay */
-    size_t samples_to_discard; /* for custom mpeg looping */
+    int skip_samples; /* base encoder delay */
+    int samples_to_discard; /* for custom mpeg looping */
 
+    int16_t* sbuf;                // decoded samples from all streams
+    int sbuf_size;              // in bytes for mpg123
 };
 
 int mpeg_custom_setup_init_default(STREAMFILE* sf, off_t start_offset, mpeg_codec_data* data, coding_t* coding_type);
